[Platinum II] Title: 나무 자르기, Time: 48 ms, Memory: 6496 KB :: cross function resolve floating point

+++ 백준/Platinum/13263. 나무 자르기/나무 자르기.c
@@ -16,7 +16,7 @@ Line* F[MAX];
 int temp = 0;
 
 double cross(Line* a, Line* b) {
-    return (b->b - a->b) / (a->a - b->a);
+    return (double)(b->b - a->b) / (a->a - b->a);
 }
 
 int lowerBound(int idx) {
